catalog-backend: moved createEntityArrayJsonStream to response utils
